♻️ (clear-signing-tester) [NO-ISSUE]: Move image inspection behind DockerContainer abstraction
Move Docker CLI image inspection logic (manifest digests, remote
labels, platform resolution) from SpeculosServiceController into
the DockerContainer interface and NodeDockerContainer adapter. The
controller now delegates all Docker operations through the injected
abstraction, making it platform-agnostic and easier to test/mock.

Made-with: Cursor
